change deci_value data type to uint8_t from uint16_t

+++ can_common.h
@@ -177,7 +177,7 @@ bool build_gps_info_msg(uint32_t timestamp,
 bool build_radi_info_msg(uint32_t timestamp,
                          uint8_t board_num,
                          uint8_t int_value,
-                         uint16_t deci_value,
+                         uint8_t deci_value,
                          can_msg_t *output);
 /*
  * Gets the general command contained in a general command message.
@@ -320,7 +320,7 @@ bool get_fill_info(const can_msg_t *msg,
 bool get_radi_info(const can_msg_t* msg,
                    uint8_t *board_num,
                    uint8_t *int_value,
-                   uint16_t *deci_value);
+                   uint8_t *deci_value);
 
 /*
  * If MSG is a DEBUG_MSG message, return its debug level, else return
